refactor(tarko-agent-ui): extract shared agent icon utility and use lightning bolt as default
